Add files via upload
A basic Python program is a simple script written in the Python programming language. It typically consists of a series of instructions that the computer executes sequentially. Python, known for its readability and simplicity, allows developers to write concise and effective code. A basic Python program might include variables, data types, conditional statements (like if-else), loops, and basic input/output operations. This foundational level of programming is crucial for beginners and serves as a stepping stone for more advanced topics in Python and other programming languages.